+++ releaf-core/lib/generators/releaf/templates/migrations/create_releaf_translations.rb
@@ -1,21 +1,19 @@
 class CreateReleafTranslations < ActiveRecord::Migration
   def change
-    create_table :releaf_translations do |t|
-      t.string  :key,       :null => false
-
-      t.timestamps(null: false)
+    create_table :releaf_i18n_entries do |t|
+      t.string :key, null: false
+      t.timestamps null: false
     end
-    add_index :releaf_translations, :key
+    add_index :releaf_i18n_entries, :key
 
-    create_table :releaf_translation_data do |t|
-      t.integer :translation_id, :null => false
-      t.string :lang, :null => false, :limit => 5
-      t.text :localization
-
-      t.timestamps(null: false)
+    create_table :releaf_i18n_entry_translations do |t|
+      t.integer :i18n_entry_id, null: false
+      t.string :locale, null: false, limit: 5
+      t.text :text
+      t.timestamps null: false
     end
-    add_index :releaf_translation_data, :lang
-    add_index :releaf_translation_data, :translation_id
-    add_index :releaf_translation_data, [:lang, :translation_id], :unique => true
+    add_index :releaf_i18n_entry_translations, :locale
+    add_index :releaf_i18n_entry_translations, :i18n_entry_id
+    add_index :releaf_i18n_entry_translations, [:locale, :i18n_entry_id], unique: true
   end
 end